FogBugz #383074-3: fpga-region: add fpga image information struct
A FPGA design contained in a FPGA image file will likely have
particulars that affect how the image is programmed to the FPGA.

This patch will allocate a struct fpga_image_info and populate
it with information from the device tree overlay that is driving
the reprogramming.  The info is then passed to the FPGA Manager
and FPGA Bridge drivers during programming.

Signed-off-by: Alan Tull <atull@opensource.altera.com>