FogBugz #383074-4: doc: fpga-mgr: add fpga image info to api
A FPGA design contained in a FPGA image file will likely have
particulars that affect how the image is programmed to the
FPGA.  Currently the only such particular is a single flag
bit indicating whether the image is for full or partial
reconfiguration.

This patch adds a new 'struct  fpga_image_info', moves the
current 'u32 flags' to it.

The FPGA Manager API functions are changed, as the 'u32 flag'
parameter is replaced with a pointer to struct fpga_image_info.

Signed-off-by: Alan Tull <atull@opensource.altera.com>