FogBugz #403733-2: completely configure l2c_aux_val
Depending on the execution path, the A10 boot rom may or
may not set some bits in the l2c aux ctrl register.  Due to this
abiguity, linux must explicitly set the register.  This patch
forces the configuration to match the full boot flow, which
also matches the setting used in the 3.10-ltsi version of the
kernel.

Signed-off-by: Matthew Gerlach <matthew.gerlach@intel.com>
---
v2:
    removed redundant EARLY_BRESP and FULL_LINE_ZERO